Problem5/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
jose==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
